+++ src/main/java/PodoeMarket/podoemarket/profile/dto/request/DetailUpdateRequestDTO.java
@@ -16,9 +16,9 @@ public class DetailUpdateRequestDTO {
     private String title;
     private String imagePath;
     private Boolean script;
-    private Integer scriptPrice;
+    private Long scriptPrice;
     private Boolean performance;
-    private Integer performancePrice;
+    private Long performancePrice;
     private String descriptionPath;
     private String plot;
